requirements.txt
streamlit 
os
librosa 
matplotlib.pyplot 